Small fix in sample project.

+++ sample/src/main/java/io/github/douglasjunior/androidSimpleTooltip/sample/MainActivity.java
@@ -166,7 +166,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                     .modal(true)
                     .animated(true)
                     .animationDuration(2000)
-                    .animationPadding((int) SimpleTooltipUtils.pxFromDp(50))
+                    .animationPadding(SimpleTooltipUtils.pxFromDp(50))
                     .contentView(R.layout.tooltip_custom, R.id.tv_text)
                     .build();
